fix overflowyness for the new rule
git-svn-id: http://svn.vuze.com/public/client@13713 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/com/aelitis/azureus/plugins/startstoprules/defaultplugin/DefaultRankCalculator.java
@@ -631,7 +631,7 @@ public class DefaultRankCalculator implements Comparable {
 					if(lastModifiedScrapeResultPeers > lastModifiedScrapeResultSeeds * 10)
 						newSR = 100 * lastModifiedScrapeResultPeers * 10;
 					else
-						newSR = 100 * lastModifiedScrapeResultPeers * lastModifiedScrapeResultPeers/(lastModifiedScrapeResultSeeds+1);
+						newSR = (int)((long)100 * lastModifiedScrapeResultPeers * lastModifiedScrapeResultPeers/(lastModifiedScrapeResultSeeds+1));
 				}				
 				else if ((iRankType == StartStopRulesDefaultPlugin.RANK_SEEDCOUNT)
 						&& (iRankTypeSeedFallback == 0 || iRankTypeSeedFallback > lastModifiedScrapeResultSeeds))
